Revert "Refactor DC Sensitivity Analysis with Woodbury engine (#1001)" (#1057)
