+++ src/main/java/com/powsybl/openloadflow/dc/DcLoadFlowEngine.java
@@ -9,7 +9,6 @@ package com.powsybl.openloadflow.dc;
 
 import com.powsybl.commons.report.ReportNode;
 import com.powsybl.loadflow.LoadFlowParameters;
-import com.powsybl.math.matrix.DenseMatrix;
 import com.powsybl.math.matrix.MatrixException;
 import com.powsybl.openloadflow.dc.equations.DcEquationType;
 import com.powsybl.openloadflow.dc.equations.DcVariableType;
@@ -154,19 +153,6 @@ public class DcLoadFlowEngine implements LoadFlowEngine<DcVariableType, DcEquati
         } catch (MatrixException e) {
             Reports.reportDcLfSolverFailure(reportNode, e.getMessage());
             LOGGER.error("Failed to solve linear system for DC load flow", e);
-            return false;
-        }
-    }
-
-    public static boolean solve(DenseMatrix targetVectorsMatrix,
-                                JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix,
-                                ReportNode reporter) {
-        try {
-            jacobianMatrix.solveTransposed(targetVectorsMatrix);
-            return true;
-        } catch (MatrixException e) {
-            Reports.reportDcLfSolverFailure(reporter, e.getMessage());
-            LOGGER.error("Failed to solve linear system for DC load flow on multiple right hand sides", e);
             return false;
         }
     }
