--- src/main/java/com/powsybl/openloadflow/dc/WoodburyEngine.java
@@ -1,112 +1,0 @@
-/**
- * Copyright (c) 2020, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
- */
-package com.powsybl.openloadflow.dc;
-
-import com.powsybl.math.matrix.DenseMatrix;
-import com.powsybl.math.matrix.LUDecomposition;
-import com.powsybl.openloadflow.dc.equations.AbstractClosedBranchDcFlowEquationTerm;
-import com.powsybl.openloadflow.dc.equations.ClosedBranchSide1DcFlowEquationTerm;
-import com.powsybl.openloadflow.dc.equations.DcEquationSystemCreationParameters;
-import com.powsybl.openloadflow.network.LfBranch;
-import com.powsybl.openloadflow.network.PiModel;
-import com.powsybl.openloadflow.network.impl.PropagatedContingency;
-
-import java.util.*;
-import java.util.function.ObjDoubleConsumer;
-
-/**
- * @author Geoffroy Jamgotchian {@literal <geoffroy.jamgotchian at rte-france.com>}
- * @author Gaël Macherel {@literal <gael.macherel@artelys.com>}
- */
-public class WoodburyEngine {
-
-    /**
-     * Compute the flow transfer factors needed to calculate post-contingency states.
-     */
-    private static void setAlphas(DcLoadFlowContext loadFlowContext, Collection<ComputedContingencyElement> contingencyElements, DenseMatrix states,
-                                  DenseMatrix contingenciesStates, int columnState, ObjDoubleConsumer<ComputedContingencyElement> setValue) {
-        if (contingencyElements.size() == 1) {
-            ComputedContingencyElement element = contingencyElements.iterator().next();
-            LfBranch lfBranch = element.getLfBranch();
-            ClosedBranchSide1DcFlowEquationTerm p1 = element.getLfBranchEquation();
-            // we solve a*alpha = b
-            double a = 1d / calculatePower(loadFlowContext, lfBranch) - (contingenciesStates.get(p1.getPh1Var().getRow(), element.getContingencyIndex())
-                    - contingenciesStates.get(p1.getPh2Var().getRow(), element.getContingencyIndex()));
-            double b = states.get(p1.getPh1Var().getRow(), columnState) - states.get(p1.getPh2Var().getRow(), columnState);
-            setValue.accept(element, b / a);
-        } else {
-            ComputedContingencyElement.setLocalIndexes(contingencyElements);
-            DenseMatrix rhs = new DenseMatrix(contingencyElements.size(), 1);
-            DenseMatrix matrix = new DenseMatrix(contingencyElements.size(), contingencyElements.size());
-            for (ComputedContingencyElement element : contingencyElements) {
-                LfBranch lfBranch = element.getLfBranch();
-                ClosedBranchSide1DcFlowEquationTerm p1 = element.getLfBranchEquation();
-                rhs.set(element.getLocalIndex(), 0, states.get(p1.getPh1Var().getRow(), columnState)
-                        - states.get(p1.getPh2Var().getRow(), columnState)
-                );
-                for (ComputedContingencyElement element2 : contingencyElements) {
-                    double value = 0d;
-                    if (element.equals(element2)) {
-                        value = 1d / calculatePower(loadFlowContext, lfBranch);
-                    }
-                    value = value - (contingenciesStates.get(p1.getPh1Var().getRow(), element2.getContingencyIndex())
-                            - contingenciesStates.get(p1.getPh2Var().getRow(), element2.getContingencyIndex()));
-                    matrix.set(element.getLocalIndex(), element2.getLocalIndex(), value);
-                }
-            }
-            try (LUDecomposition lu = matrix.decomposeLU()) {
-                lu.solve(rhs); // rhs now contains state matrix
-            }
-            contingencyElements.forEach(element -> setValue.accept(element, rhs.get(element.getLocalIndex(), 0)));
-        }
-    }
-
-    private static double calculatePower(DcLoadFlowContext loadFlowContext, LfBranch lfBranch) {
-        PiModel piModel = lfBranch.getPiModel();
-        DcEquationSystemCreationParameters creationParameters = loadFlowContext.getParameters().getEquationSystemCreationParameters();
-        return AbstractClosedBranchDcFlowEquationTerm.calculatePower(creationParameters.isUseTransformerRatio(), creationParameters.getDcApproximationType(), piModel);
-    }
-
-    /**
-     * Calculate post-contingency states using pre-contingency angle values and some flow transfer factors (alphas).
-     */
-    private DenseMatrix computePostContingencyStates(DcLoadFlowContext loadFlowContext, DenseMatrix preContingencyStates, DenseMatrix contingenciesStates, Collection<ComputedContingencyElement> contingencyElements) {
-        // fill the post contingency matrices
-        DenseMatrix postContingencyStates = new DenseMatrix(preContingencyStates.getRowCount(), preContingencyStates.getColumnCount());
-        for (int columnIndex = 0; columnIndex < preContingencyStates.getColumnCount(); columnIndex++) {
-            setAlphas(loadFlowContext, contingencyElements, preContingencyStates, contingenciesStates, columnIndex, ComputedContingencyElement::setAlphaForPostContingencyState);
-            for (int rowIndex = 0; rowIndex < preContingencyStates.getRowCount(); rowIndex++) {
-                double postContingencyValue = preContingencyStates.get(rowIndex, columnIndex);
-                for (ComputedContingencyElement contingencyElement : contingencyElements) {
-                    postContingencyValue += contingencyElement.getAlphaForPostContingencyState() * contingenciesStates.get(rowIndex, contingencyElement.getContingencyIndex());
-                }
-                postContingencyStates.set(rowIndex, columnIndex, postContingencyValue);
-            }
-        }
-        return postContingencyStates;
-    }
-
-    /**
-     * Compute post-contingency states of a network, using Woodbury formula, and for a given {@link WoodburyEngineInputReader}.
-     * This reader provides the contingency elements and the pre-contingency angle values on which is applied the formula.
-     *
-     * @return a map with post-contingency angle states, indexed by contingencies.
-     */
-    public Map<PropagatedContingency, DenseMatrix> run(DcLoadFlowContext loadFlowContext, WoodburyEngineInputReader reader, DenseMatrix contingenciesStates) {
-        Objects.requireNonNull(loadFlowContext);
-        Objects.requireNonNull(reader);
-        Objects.requireNonNull(contingenciesStates);
-
-        Map<PropagatedContingency, DenseMatrix> postContingencyStatesByContingency = new HashMap<>();
-        reader.process((PropagatedContingency contingency, Collection<ComputedContingencyElement> contingencyElements, DenseMatrix preContingencyStates) -> {
-            DenseMatrix postContingencyStates = computePostContingencyStates(loadFlowContext, preContingencyStates, contingenciesStates, contingencyElements);
-            postContingencyStatesByContingency.put(contingency, postContingencyStates);
-        });
-        return postContingencyStatesByContingency;
-    }
-}
