+++ src/main/java/com/powsybl/openloadflow/sensi/DcSensitivityAnalysis.java
@@ -10,17 +10,23 @@ package com.powsybl.openloadflow.sensi;
 import com.google.common.base.Stopwatch;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.report.ReportNode;
+import com.powsybl.contingency.BranchContingency;
+import com.powsybl.contingency.ContingencyElement;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.DenseMatrix;
+import com.powsybl.math.matrix.LUDecomposition;
+import com.powsybl.math.matrix.Matrix;
 import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
-import com.powsybl.openloadflow.dc.*;
-import com.powsybl.openloadflow.dc.equations.DcEquationSystemCreationParameters;
-import com.powsybl.openloadflow.dc.equations.DcEquationType;
-import com.powsybl.openloadflow.dc.equations.DcVariableType;
+import com.powsybl.openloadflow.dc.DcLoadFlowContext;
+import com.powsybl.openloadflow.dc.DcLoadFlowEngine;
+import com.powsybl.openloadflow.dc.DcLoadFlowParameters;
+import com.powsybl.openloadflow.dc.equations.*;
 import com.powsybl.openloadflow.equations.Equation;
+import com.powsybl.openloadflow.equations.EquationSystem;
 import com.powsybl.openloadflow.equations.StateVector;
+import com.powsybl.openloadflow.graph.GraphConnectivity;
 import com.powsybl.openloadflow.graph.GraphConnectivityFactory;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.network.impl.LfNetworkList;
@@ -37,9 +43,9 @@ import org.apache.commons.lang3.tuple.Pair;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
+import java.util.function.ObjDoubleConsumer;
 import java.util.stream.Collectors;
 
-import static com.powsybl.openloadflow.dc.DcLoadFlowEngine.solve;
 import static com.powsybl.openloadflow.network.util.ParticipatingElement.normalizeParticipationFactors;
 
 /**
@@ -47,7 +53,168 @@ import static com.powsybl.openloadflow.network.util.ParticipatingElement.normali
  * @author Gaël Macherel {@literal <gael.macherel@artelys.com>}
  */
 public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariableType, DcEquationType> {
+
+    private static final double CONNECTIVITY_LOSS_THRESHOLD = 10e-7;
     private static final double FUNCTION_REFERENCE_ZER0_THRESHOLD = 1e-13;
+
+    static final class ComputedContingencyElement {
+
+        private int contingencyIndex = -1; // index of the element in the rhs for +1-1
+        private int localIndex = -1; // local index of the element : index of the element in the matrix used in the setAlphas method
+        private double alphaForSensitivityValue = Double.NaN;
+        private double alphaForFunctionReference = Double.NaN;
+        private final ContingencyElement element;
+        private final LfBranch lfBranch;
+        private final ClosedBranchSide1DcFlowEquationTerm branchEquation;
+
+        private ComputedContingencyElement(final ContingencyElement element, LfNetwork lfNetwork, EquationSystem<DcVariableType, DcEquationType> equationSystem) {
+            this.element = element;
+            lfBranch = lfNetwork.getBranchById(element.getId());
+            branchEquation = equationSystem.getEquationTerm(ElementType.BRANCH, lfBranch.getNum(), ClosedBranchSide1DcFlowEquationTerm.class);
+        }
+
+        private int getContingencyIndex() {
+            return contingencyIndex;
+        }
+
+        private void setContingencyIndex(final int index) {
+            this.contingencyIndex = index;
+        }
+
+        private int getLocalIndex() {
+            return localIndex;
+        }
+
+        private void setLocalIndex(final int index) {
+            this.localIndex = index;
+        }
+
+        private double getAlphaForSensitivityValue() {
+            return alphaForSensitivityValue;
+        }
+
+        private void setAlphaForSensitivityValue(final double alpha) {
+            this.alphaForSensitivityValue = alpha;
+        }
+
+        private double getAlphaForFunctionReference() {
+            return alphaForFunctionReference;
+        }
+
+        private void setAlphaForFunctionReference(final double alpha) {
+            this.alphaForFunctionReference = alpha;
+        }
+
+        private ContingencyElement getElement() {
+            return element;
+        }
+
+        private LfBranch getLfBranch() {
+            return lfBranch;
+        }
+
+        private ClosedBranchSide1DcFlowEquationTerm getLfBranchEquation() {
+            return branchEquation;
+        }
+
+        private static void setContingencyIndexes(Collection<ComputedContingencyElement> elements) {
+            int index = 0;
+            for (ComputedContingencyElement element : elements) {
+                element.setContingencyIndex(index++);
+            }
+        }
+
+        private static void setLocalIndexes(Collection<ComputedContingencyElement> elements) {
+            int index = 0;
+            for (ComputedContingencyElement element : elements) {
+                element.setLocalIndex(index++);
+            }
+        }
+
+        static void applyToConnectivity(LfNetwork lfNetwork, GraphConnectivity<LfBus, LfBranch> connectivity, Collection<ComputedContingencyElement> breakingConnectivityElements) {
+            breakingConnectivityElements.stream()
+                    .map(ComputedContingencyElement::getElement)
+                    .map(ContingencyElement::getId)
+                    .distinct()
+                    .map(lfNetwork::getBranchById)
+                    .filter(b -> b.getBus1() != null && b.getBus2() != null)
+                    .forEach(connectivity::removeEdge);
+        }
+    }
+
+    private static final class PhaseTapChangerContingenciesIndexing {
+
+        private final List<PropagatedContingency> contingenciesWithoutTransformers = new ArrayList<>();
+        private final Map<Set<LfBranch>, Collection<PropagatedContingency>> contingenciesIndexedByPhaseTapChangers = new LinkedHashMap<>();
+
+        private PhaseTapChangerContingenciesIndexing(Collection<PropagatedContingency> contingencies,
+                                                     Map<String, ComputedContingencyElement> contingencyElementByBranch,
+                                                     Collection<String> elementIdsToSkip) {
+            for (PropagatedContingency contingency : contingencies) {
+                Set<LfBranch> lostTransformers = contingency.getBranchIdsToOpen().keySet().stream()
+                        .filter(element -> !elementIdsToSkip.contains(element))
+                        .map(contingencyElementByBranch::get)
+                        .map(ComputedContingencyElement::getLfBranch)
+                        .filter(LfBranch::hasPhaseControllerCapability)
+                        .collect(Collectors.toSet());
+                if (lostTransformers.isEmpty()) {
+                    contingenciesWithoutTransformers.add(contingency);
+                } else {
+                    contingenciesIndexedByPhaseTapChangers.computeIfAbsent(lostTransformers, key -> new ArrayList<>()).add(contingency);
+                }
+            }
+        }
+
+        private Collection<PropagatedContingency> getContingenciesWithoutPhaseTapChangerLoss() {
+            return contingenciesWithoutTransformers;
+        }
+
+        private Map<Set<LfBranch>, Collection<PropagatedContingency>> getContingenciesIndexedByPhaseTapChangers() {
+            return contingenciesIndexedByPhaseTapChangers;
+        }
+    }
+
+    private static final class ConnectivityAnalysisResult {
+
+        private final Collection<PropagatedContingency> contingencies = new HashSet<>();
+
+        private final Set<String> elementsToReconnect;
+
+        private final Set<LfBus> disabledBuses;
+
+        private final Set<LfBus> slackConnectedComponent;
+
+        private final Set<LfBranch> partialDisabledBranches; // branches disabled because of connectivity loss.
+
+        private ConnectivityAnalysisResult(Set<String> elementsToReconnect,
+                                           GraphConnectivity<LfBus, LfBranch> connectivity,
+                                           LfNetwork lfNetwork) {
+            this.elementsToReconnect = elementsToReconnect;
+            slackConnectedComponent = connectivity.getConnectedComponent(lfNetwork.getSlackBus());
+            disabledBuses = connectivity.getVerticesRemovedFromMainComponent();
+            partialDisabledBranches = connectivity.getEdgesRemovedFromMainComponent();
+        }
+
+        private Collection<PropagatedContingency> getContingencies() {
+            return contingencies;
+        }
+
+        private Set<String> getElementsToReconnect() {
+            return elementsToReconnect;
+        }
+
+        private Set<LfBus> getDisabledBuses() {
+            return disabledBuses;
+        }
+
+        private Set<LfBus> getSlackConnectedComponent() {
+            return slackConnectedComponent;
+        }
+
+        private Set<LfBranch> getPartialDisabledBranches() {
+            return partialDisabledBranches;
+        }
+    }
 
     public DcSensitivityAnalysis(MatrixFactory matrixFactory, GraphConnectivityFactory<LfBus, LfBranch> connectivityFactory, SensitivityAnalysisParameters parameters) {
         super(matrixFactory, connectivityFactory, parameters);
@@ -71,123 +238,14 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
                 .setMaxOuterLoopIterations(parametersExt.getMaxOuterLoopIterations());
     }
 
-    private void createBranchPreContingencySensitivityValue(LfSensitivityFactor<DcVariableType, DcEquationType> factor, SensitivityResultWriter resultWriter) {
-        Pair<Optional<Double>, Optional<Double>> predefinedResults = getPredefinedResults(factor, new DisabledNetwork(), null);
-        Optional<Double> sensitivityValuePredefinedResult = predefinedResults.getLeft();
-        Optional<Double> functionPredefinedResults = predefinedResults.getRight();
-        double sensitivityValue = sensitivityValuePredefinedResult.orElseGet(factor::getBaseSensitivityValue);
-        double functionValue = functionPredefinedResults.orElseGet(factor::getFunctionReference);
-        double unscaledSensi = unscaleSensitivity(factor, sensitivityValue);
-        if (!filterSensitivityValue(unscaledSensi, factor.getVariableType(), factor.getFunctionType(), parameters)) {
-            resultWriter.writeSensitivityValue(factor.getIndex(), -1, unscaledSensi, unscaleFunction(factor, functionValue));
-        }
-    }
-
-    private void createBranchPostContingenciesSensitivityValue(LfSensitivityFactor<DcVariableType, DcEquationType> factor, SensitivityFactorGroup<DcVariableType, DcEquationType> factorGroup,
-                                                               List<PropagatedContingency> contingencies, Map<PropagatedContingency, DenseMatrix> injectionResult, Map<PropagatedContingency, DenseMatrix> flowResult,
-                                                               HashMap<PropagatedContingency, DisabledNetwork> disabledNetworksByPropagatedContingencies,
-                                                               SensitivityResultWriter resultWriter) {
-        Derivable<DcVariableType> p1 = factor.getFunctionEquationTerm();
-        for (PropagatedContingency contingency : contingencies) {
-            DenseMatrix postContingencyInjectionStates = injectionResult.get(contingency);
-            DenseMatrix postContingencyFlowStates = flowResult.get(contingency);
-            DisabledNetwork disabledNetwork = disabledNetworksByPropagatedContingencies.get(contingency);
-
-            Pair<Optional<Double>, Optional<Double>> predefinedResults = getPredefinedResults(factor, disabledNetwork, contingency);
-            Optional<Double> sensitivityValuePredefinedResult = predefinedResults.getLeft();
-            Optional<Double> functionPredefinedResults = predefinedResults.getRight();
-
-            double sensitivityValue = sensitivityValuePredefinedResult.orElseGet(factor::getBaseSensitivityValue);
-            double functionValue = functionPredefinedResults.orElseGet(factor::getFunctionReference);
-
-            if (sensitivityValuePredefinedResult.isEmpty()) {
-                sensitivityValue = p1.calculateSensi(postContingencyInjectionStates, factorGroup.getIndex());
-            }
-
-            if (functionPredefinedResults.isEmpty()) {
-                functionValue = p1.calculateSensi(postContingencyFlowStates, 0);
-            }
-
-            functionValue = fixZeroFunctionReference(contingency, functionValue);
-            double unscaledSensi = unscaleSensitivity(factor, sensitivityValue);
-            if (!filterSensitivityValue(unscaledSensi, factor.getVariableType(), factor.getFunctionType(), parameters)) {
-                int contingencyIndex = contingency != null ? contingency.getIndex() : -1;
-                resultWriter.writeSensitivityValue(factor.getIndex(), contingencyIndex, unscaledSensi, unscaleFunction(factor, functionValue));
-            }
-        }
-    }
-
     /**
-     * Post contingency reference flow, that should be strictly zero, for numeric reason and because it is computed
-     * from shifting pre-contingency non-zero flow, cannot end up to a strict zero: very small values are converted to zero.
+     * Calculate the active power flows for pre-contingency or a post-contingency state and set the factor function reference.
+     * The interesting disabled branches are only phase shifters.
      */
-    private static double fixZeroFunctionReference(PropagatedContingency contingency, double functionValue) {
-        if (contingency != null) {
-            return Math.abs(functionValue) < FUNCTION_REFERENCE_ZER0_THRESHOLD ? 0 : functionValue;
-        }
-        return functionValue;
-    }
-
-    /**
-     * Calculate the sensitivity value for pre-contingency state only.
-     */
-    private void setBaseCaseSensitivityValues(SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups, DenseMatrix factorsState) {
-        for (SensitivityFactorGroup<DcVariableType, DcEquationType> factorGroup : factorGroups.getList()) {
-            for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factorGroup.getFactors()) {
-                factor.setBaseCaseSensitivityValue(factor.getFunctionEquationTerm().calculateSensi(factorsState, factorGroup.getIndex()));
-            }
-        }
-    }
-
-    /**
-     * Set the function reference of factors with calculated pre-contingency or post-contingency states.
-     */
-    private void setFunctionReference(List<LfSensitivityFactor<DcVariableType, DcEquationType>> factors, DenseMatrix states) {
-        double[] statesArray = new double[states.getRowCount()];
-        for (int i = 0; i < states.getRowCount(); i++) {
-            statesArray[i] = states.get(i, 0);
-        }
-        StateVector sv = new StateVector(statesArray);
-        for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factors) {
-            factor.setFunctionReference(factor.getFunctionEquationTerm().eval(sv)); // pass explicitly the previously calculated state vector
-        }
-    }
-
-    /**
-     * Calculate sensitivity values for post-contingency state.
-     */
-    private void calculateSensitivityValues(Map<PropagatedContingency, DenseMatrix> injectionResult, Map<PropagatedContingency, DenseMatrix> flowResult, HashMap<PropagatedContingency, DisabledNetwork> disabledNetworksByPropagatedContingencies,
-                                            List<LfSensitivityFactor<DcVariableType, DcEquationType>> lfFactors, List<PropagatedContingency> contingencies,
-                                            SensitivityResultWriter resultWriter) {
-        if (lfFactors.isEmpty()) {
-            return;
-        }
-
-        lfFactors.stream().filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID_ONLY_FOR_FUNCTION)
-                .forEach(factor -> {
-                    createBranchPreContingencySensitivityValue(factor, resultWriter);
-                    createBranchPostContingenciesSensitivityValue(factor, null, contingencies, injectionResult, flowResult, disabledNetworksByPropagatedContingencies, resultWriter);
-                });
-
-        Map<SensitivityFactorGroup<DcVariableType, DcEquationType>, List<LfSensitivityFactor<DcVariableType, DcEquationType>>> factorsByGroup = lfFactors.stream()
-                .filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID)
-                .collect(Collectors.groupingBy(LfSensitivityFactor::getGroup, LinkedHashMap::new, Collectors.toList()));
-        for (Map.Entry<SensitivityFactorGroup<DcVariableType, DcEquationType>, List<LfSensitivityFactor<DcVariableType, DcEquationType>>> e : factorsByGroup.entrySet()) {
-            SensitivityFactorGroup<DcVariableType, DcEquationType> factorGroup = e.getKey();
-            List<LfSensitivityFactor<DcVariableType, DcEquationType>> factorsForThisGroup = e.getValue();
-            for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factorsForThisGroup) {
-                createBranchPreContingencySensitivityValue(factor, resultWriter);
-                createBranchPostContingenciesSensitivityValue(factor, factorGroup, contingencies, injectionResult, flowResult, disabledNetworksByPropagatedContingencies, resultWriter);
-            }
-        }
-    }
-
-    /**
-     * Compute flow rhs taking into account slack distribution.
-     */
-    private static DenseMatrix getPreContingencyFlowRhs(DcLoadFlowContext loadFlowContext,
-                                                        List<ParticipatingElement> participatingElements,
-                                                        DisabledNetwork disabledNetwork) {
+    private DenseMatrix calculateActivePowerFlows(DcLoadFlowContext loadFlowContext,
+                                                  List<LfSensitivityFactor<DcVariableType, DcEquationType>> factors, List<ParticipatingElement> participatingElements,
+                                                  DisabledNetwork disabledNetwork,
+                                                  ReportNode reportNode) {
         List<BusState> busStates = Collections.emptyList();
         DcLoadFlowParameters parameters = loadFlowContext.getParameters();
         if (parameters.isDistributedSlack()) {
@@ -196,21 +254,26 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
                     .collect(Collectors.toSet()), BusState::save);
         }
 
-        double[] preContingencyFlowRhs = getDcLoadFlowTargetVector(loadFlowContext, disabledNetwork);
+        double[] dx = runDcLoadFlow(loadFlowContext, disabledNetwork, reportNode);
+
+        StateVector sv = new StateVector(dx);
+        for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factors) {
+            factor.setFunctionReference(factor.getFunctionEquationTerm().eval(sv)); // pass explicitly the previously calculated state vector
+        }
 
         if (parameters.isDistributedSlack()) {
             ElementState.restore(busStates);
         }
 
-        return new DenseMatrix(preContingencyFlowRhs.length, 1, preContingencyFlowRhs);
+        return new DenseMatrix(dx.length, 1, dx);
     }
 
     /**
-     * A simplified version of DcLoadFlowEngine that supports on the fly bus and branch disabling, that only
-     * returns the target vector of the equation system.
+     * A simplified version of DcLoadFlowEngine that supports on the fly bus and branch disabling and that do not
+     * update the state vector and the network at the end (because we don't need it to just evaluate a few equations)
      */
-    private static double[] getDcLoadFlowTargetVector(DcLoadFlowContext loadFlowContext, DisabledNetwork disabledNetwork) {
-
+    public double[] runDcLoadFlow(DcLoadFlowContext loadFlowContext, DisabledNetwork disabledNetwork,
+                                  ReportNode reportNode) {
         Collection<LfBus> remainingBuses;
         if (disabledNetwork.getBuses().isEmpty()) {
             remainingBuses = loadFlowContext.getNetwork().getBuses();
@@ -247,46 +310,568 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
                     .forEach(column -> targetVectorArray[column] = 0);
         }
 
-        return targetVectorArray;
+        boolean succeeded = DcLoadFlowEngine.solve(targetVectorArray, loadFlowContext.getJacobianMatrix(), reportNode);
+        if (!succeeded) {
+            throw new PowsyblException("DC solver failed");
+        }
+
+        return targetVectorArray; // now contains dx
     }
 
     /**
-     * Compute all the injection rhs taking into account slack distribution.
+     * Get the sensitivity value for pre-contingency state and calculate the sensitivity value for a post-contingency state if asked.
+     * The sensitivity value is written in the SensitivityResultWriter.
      */
-    static DenseMatrix getPreContingencyInjectionRhs(DcLoadFlowContext loadFlowContext,
-                                                     SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups,
-                                                     List<ParticipatingElement> participatingElements) {
-        Map<LfBus, Double> slackParticipationByBus;
-        if (participatingElements.isEmpty()) {
-            slackParticipationByBus = new HashMap<>();
-        } else {
-            slackParticipationByBus = participatingElements.stream().collect(Collectors.toMap(
-                    ParticipatingElement::getLfBus,
-                    element -> -element.getFactor(),
-                    Double::sum));
+    private void createBranchSensitivityValue(LfSensitivityFactor<DcVariableType, DcEquationType> factor, DenseMatrix contingenciesStates,
+                                              Collection<ComputedContingencyElement> contingencyElements,
+                                              PropagatedContingency contingency, SensitivityResultWriter resultWriter,
+                                              DisabledNetwork disabledNetwork) {
+        Pair<Optional<Double>, Optional<Double>> predefinedResults = getPredefinedResults(factor, disabledNetwork, contingency);
+        Optional<Double> sensitivityValuePredefinedResult = predefinedResults.getLeft();
+        Optional<Double> functionPredefinedResults = predefinedResults.getRight();
+        double sensitivityValue = sensitivityValuePredefinedResult.orElseGet(factor::getBaseSensitivityValue);
+        double functionValue = functionPredefinedResults.orElseGet(factor::getFunctionReference);
+        Derivable<DcVariableType> p1 = factor.getFunctionEquationTerm();
+
+        if (!(functionPredefinedResults.isPresent() && sensitivityValuePredefinedResult.isPresent())) {
+            for (ComputedContingencyElement contingencyElement : contingencyElements) {
+                double contingencySensitivity = p1.calculateSensi(contingenciesStates, contingencyElement.getContingencyIndex());
+                if (functionPredefinedResults.isEmpty()) {
+                    functionValue += contingencyElement.getAlphaForFunctionReference() * contingencySensitivity;
+                }
+                if (sensitivityValuePredefinedResult.isEmpty()) {
+                    sensitivityValue += contingencyElement.getAlphaForSensitivityValue() * contingencySensitivity;
+                }
+            }
         }
 
-        return initFactorsRhs(loadFlowContext.getEquationSystem(), factorGroups, slackParticipationByBus);
+        functionValue = fixZeroFunctionReference(contingency, functionValue);
+
+        double unscaledSensi = unscaleSensitivity(factor, sensitivityValue);
+        if (!filterSensitivityValue(unscaledSensi, factor.getVariableType(), factor.getFunctionType(), parameters)) {
+            resultWriter.writeSensitivityValue(factor.getIndex(), contingency != null ? contingency.getIndex() : -1, unscaledSensi, unscaleFunction(factor, functionValue));
+        }
     }
 
-    private List<ParticipatingElement> getNewNormalizedParticipationFactors(DcLoadFlowContext loadFlowContext, OpenLoadFlowParameters lfParametersExt,
-                                                                            LfContingency lfContingency, List<ParticipatingElement> participatingElements) {
-        LfNetwork lfNetwork = loadFlowContext.getNetwork();
-        DcLoadFlowParameters lfParameters = loadFlowContext.getParameters();
-        List<ParticipatingElement> newParticipatingElements;
-        if (isDistributedSlackOnGenerators(loadFlowContext.getParameters())) {
-            // deep copy of participatingElements, removing the participating LfGeneratorImpl whose targetP has been set to 0
-            Set<LfGenerator> participatingGeneratorsToRemove = lfContingency.getLostGenerators();
-            newParticipatingElements = participatingElements.stream()
-                    .filter(participatingElement -> !(participatingElement.getElement() instanceof LfGenerator lfGenerator
-                            && participatingGeneratorsToRemove.contains(lfGenerator)))
-                    .map(participatingElement -> new ParticipatingElement(participatingElement.getElement(), participatingElement.getFactor()))
-                    .toList();
-            normalizeParticipationFactors(newParticipatingElements);
-        } else { // slack distribution on loads
-            newParticipatingElements = getParticipatingElements(lfNetwork.getBuses(), lfParameters.getBalanceType(), lfParametersExt);
+    /**
+     * Post contingency reference flow, that should be strictly zero, for numeric reason and because it is computed
+     * from shifting pre-contingency non-zero flow, cannot end up to a strict zero: very small values are converted to zero.
+     */
+    private static double fixZeroFunctionReference(PropagatedContingency contingency, double functionValue) {
+        if (contingency != null) {
+            return Math.abs(functionValue) < FUNCTION_REFERENCE_ZER0_THRESHOLD ? 0 : functionValue;
         }
-        return newParticipatingElements;
+        return functionValue;
+    }
+
+    /**
+     * Calculate the sensitivity value for pre-contingency state only.
+     */
+    private void setBaseCaseSensitivityValues(SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups, DenseMatrix factorsState) {
+        for (SensitivityFactorGroup<DcVariableType, DcEquationType> factorGroup : factorGroups.getList()) {
+            for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factorGroup.getFactors()) {
+                factor.setBaseCaseSensitivityValue(factor.getFunctionEquationTerm().calculateSensi(factorsState, factorGroup.getIndex()));
+            }
+        }
+    }
+
+    /**
+     * Compute state for sensitivity factors taking into account slack distribution.
+     */
+    private DenseMatrix calculateFactorStates(DcLoadFlowContext loadFlowContext,
+                                              SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups,
+                                              List<ParticipatingElement> participatingElements) {
+        Map<LfBus, Double> slackParticipationByBus;
+        if (participatingElements.isEmpty()) {
+            slackParticipationByBus = Map.of(loadFlowContext.getNetwork().getSlackBus(), -1d);
+        } else {
+            slackParticipationByBus = participatingElements.stream().collect(Collectors.toMap(
+                ParticipatingElement::getLfBus,
+                element -> -element.getFactor(),
+                Double::sum));
+        }
+
+        DenseMatrix factorStates = initFactorsRhs(loadFlowContext.getEquationSystem(), factorGroups, slackParticipationByBus);
+        loadFlowContext.getJacobianMatrix().solveTransposed(factorStates); // states for the sensitivity factors
+        setBaseCaseSensitivityValues(factorGroups, factorStates); // use this state to compute the base sensitivity (without +1-1)
+        return factorStates;
+    }
+
+    /**
+     * Calculate sensitivity values for pre-contingency state or a post-contingency state using the pre-contingency sensitivity
+     * value and some flow transfer factors (alphas).
+     */
+    private void calculateSensitivityValues(DcLoadFlowContext loadFlowContext, List<LfSensitivityFactor<DcVariableType, DcEquationType>> lfFactors, DenseMatrix factorStates,
+                                            DenseMatrix contingenciesStates, DenseMatrix flowStates, Collection<ComputedContingencyElement> contingencyElements,
+                                            PropagatedContingency contingency, SensitivityResultWriter resultWriter, DisabledNetwork disabledNetwork) {
+        if (lfFactors.isEmpty()) {
+            return;
+        }
+
+        setAlphas(loadFlowContext, contingencyElements, flowStates, contingenciesStates, 0, ComputedContingencyElement::setAlphaForFunctionReference);
+
+        lfFactors.stream().filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID_ONLY_FOR_FUNCTION)
+                .forEach(factor -> createBranchSensitivityValue(factor, contingenciesStates, contingencyElements, contingency, resultWriter, disabledNetwork));
+
+        Map<SensitivityFactorGroup<DcVariableType, DcEquationType>, List<LfSensitivityFactor<DcVariableType, DcEquationType>>> factorsByGroup = lfFactors.stream()
+                .filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID)
+                .collect(Collectors.groupingBy(LfSensitivityFactor::getGroup, LinkedHashMap::new, Collectors.toList()));
+        for (Map.Entry<SensitivityFactorGroup<DcVariableType, DcEquationType>, List<LfSensitivityFactor<DcVariableType, DcEquationType>>> e : factorsByGroup.entrySet()) {
+            SensitivityFactorGroup<DcVariableType, DcEquationType> factorGroup = e.getKey();
+            List<LfSensitivityFactor<DcVariableType, DcEquationType>> factorsForThisGroup = e.getValue();
+            setAlphas(loadFlowContext, contingencyElements, factorStates, contingenciesStates, factorGroup.getIndex(), ComputedContingencyElement::setAlphaForSensitivityValue);
+            for (LfSensitivityFactor<DcVariableType, DcEquationType> factor : factorsForThisGroup) {
+                createBranchSensitivityValue(factor, contingenciesStates, contingencyElements, contingency, resultWriter, disabledNetwork);
+            }
+        }
+    }
+
+    /**
+     * Compute the flow transfer factors needed to calculate the post-contingency sensitivity values.
+     */
+    private static void setAlphas(DcLoadFlowContext loadFlowContext, Collection<ComputedContingencyElement> contingencyElements, DenseMatrix states,
+                                  DenseMatrix contingenciesStates, int columnState, ObjDoubleConsumer<ComputedContingencyElement> setValue) {
+        if (contingencyElements.size() == 1) {
+            ComputedContingencyElement element = contingencyElements.iterator().next();
+            LfBranch lfBranch = element.getLfBranch();
+            ClosedBranchSide1DcFlowEquationTerm p1 = element.getLfBranchEquation();
+            // we solve a*alpha = b
+            double a = 1d / calculatePower(loadFlowContext, lfBranch) - (contingenciesStates.get(p1.getPh1Var().getRow(), element.getContingencyIndex())
+                    - contingenciesStates.get(p1.getPh2Var().getRow(), element.getContingencyIndex()));
+            double b = states.get(p1.getPh1Var().getRow(), columnState) - states.get(p1.getPh2Var().getRow(), columnState);
+            setValue.accept(element, b / a);
+        } else {
+            // FIXME: direct resolution if contingencyElements.size() == 2
+            ComputedContingencyElement.setLocalIndexes(contingencyElements);
+            DenseMatrix rhs = new DenseMatrix(contingencyElements.size(), 1);
+            DenseMatrix matrix = new DenseMatrix(contingencyElements.size(), contingencyElements.size());
+            for (ComputedContingencyElement element : contingencyElements) {
+                LfBranch lfBranch = element.getLfBranch();
+                ClosedBranchSide1DcFlowEquationTerm p1 = element.getLfBranchEquation();
+                rhs.set(element.getLocalIndex(), 0, states.get(p1.getPh1Var().getRow(), columnState)
+                        - states.get(p1.getPh2Var().getRow(), columnState)
+                );
+                for (ComputedContingencyElement element2 : contingencyElements) {
+                    double value = 0d;
+                    if (element.equals(element2)) {
+                        value = 1d / calculatePower(loadFlowContext, lfBranch);
+                    }
+                    value = value - (contingenciesStates.get(p1.getPh1Var().getRow(), element2.getContingencyIndex())
+                            - contingenciesStates.get(p1.getPh2Var().getRow(), element2.getContingencyIndex()));
+                    matrix.set(element.getLocalIndex(), element2.getLocalIndex(), value);
+                }
+            }
+            try (LUDecomposition lu = matrix.decomposeLU()) {
+                lu.solve(rhs); // rhs now contains state matrix
+            }
+            contingencyElements.forEach(element -> setValue.accept(element, rhs.get(element.getLocalIndex(), 0)));
+        }
+    }
+
+    private static double calculatePower(DcLoadFlowContext loadFlowContext, LfBranch lfBranch) {
+        PiModel piModel = lfBranch.getPiModel();
+        DcEquationSystemCreationParameters creationParameters = loadFlowContext.getParameters().getEquationSystemCreationParameters();
+        return AbstractClosedBranchDcFlowEquationTerm.calculatePower(creationParameters.isUseTransformerRatio(), creationParameters.getDcApproximationType(), piModel);
+    }
+
+    /**
+     * Fills the right hand side with +1/-1 to model a branch contingency.
+     */
+    private static void fillRhsContingency(LfNetwork lfNetwork, EquationSystem<DcVariableType, DcEquationType> equationSystem,
+                                           Collection<ComputedContingencyElement> contingencyElements, Matrix rhs) {
+        for (ComputedContingencyElement element : contingencyElements) {
+            LfBranch lfBranch = lfNetwork.getBranchById(element.getElement().getId());
+            if (lfBranch.getBus1() == null || lfBranch.getBus2() == null) {
+                continue;
+            }
+            LfBus bus1 = lfBranch.getBus1();
+            LfBus bus2 = lfBranch.getBus2();
+            if (bus1.isSlack()) {
+                Equation<DcVariableType, DcEquationType> p = equationSystem.getEquation(bus2.getNum(), DcEquationType.BUS_TARGET_P).orElseThrow(IllegalStateException::new);
+                rhs.set(p.getColumn(), element.getContingencyIndex(), -1);
+            } else if (bus2.isSlack()) {
+                Equation<DcVariableType, DcEquationType> p = equationSystem.getEquation(bus1.getNum(), DcEquationType.BUS_TARGET_P).orElseThrow(IllegalStateException::new);
+                rhs.set(p.getColumn(), element.getContingencyIndex(), 1);
+            } else {
+                Equation<DcVariableType, DcEquationType> p1 = equationSystem.getEquation(bus1.getNum(), DcEquationType.BUS_TARGET_P).orElseThrow(IllegalStateException::new);
+                Equation<DcVariableType, DcEquationType> p2 = equationSystem.getEquation(bus2.getNum(), DcEquationType.BUS_TARGET_P).orElseThrow(IllegalStateException::new);
+                rhs.set(p1.getColumn(), element.getContingencyIndex(), 1);
+                rhs.set(p2.getColumn(), element.getContingencyIndex(), -1);
+            }
+        }
+    }
+
+    static DenseMatrix initContingencyRhs(LfNetwork lfNetwork, EquationSystem<DcVariableType, DcEquationType> equationSystem, Collection<ComputedContingencyElement> contingencyElements) {
+        // otherwise, defining the rhs matrix will result in integer overflow
+        int equationCount = equationSystem.getIndex().getSortedEquationsToSolve().size();
+        int maxContingencyElements = Integer.MAX_VALUE / (equationCount * Double.BYTES);
+        if (contingencyElements.size() > maxContingencyElements) {
+            throw new PowsyblException("Too many contingency elements " + contingencyElements.size()
+                    + ", maximum is " + maxContingencyElements + " for a system with " + equationCount + " equations");
+        }
+
+        DenseMatrix rhs = new DenseMatrix(equationCount, contingencyElements.size());
+        fillRhsContingency(lfNetwork, equationSystem, contingencyElements, rhs);
+        return rhs;
+    }
+
+    private static DenseMatrix calculateContingenciesStates(DcLoadFlowContext loadFlowContext, Map<String, ComputedContingencyElement> contingencyElementByBranch) {
+        DenseMatrix contingenciesStates = initContingencyRhs(loadFlowContext.getNetwork(), loadFlowContext.getEquationSystem(), contingencyElementByBranch.values()); // rhs with +1 -1 on contingency elements
+        loadFlowContext.getJacobianMatrix().solveTransposed(contingenciesStates);
+        return contingenciesStates;
+    }
+
+    private static void detectPotentialConnectivityBreak(LfNetwork lfNetwork, DenseMatrix states, List<PropagatedContingency> contingencies,
+                                                         Map<String, ComputedContingencyElement> contingencyElementByBranch,
+                                                         EquationSystem<DcVariableType, DcEquationType> equationSystem,
+                                                         Collection<PropagatedContingency> nonLosingConnectivityContingencies,
+                                                         Map<Set<ComputedContingencyElement>, List<PropagatedContingency>> contingenciesByGroupOfElementsBreakingConnectivity) {
+        for (PropagatedContingency contingency : contingencies) {
+            List<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream().map(contingencyElementByBranch::get).collect(Collectors.toList());
+            Set<ComputedContingencyElement> groupOfElementsBreakingConnectivity = getGroupOfElementsBreakingConnectivity(lfNetwork, states, contingencyElements, equationSystem);
+            if (groupOfElementsBreakingConnectivity.isEmpty()) { // connectivity not broken
+                nonLosingConnectivityContingencies.add(contingency);
+            } else {
+                contingenciesByGroupOfElementsBreakingConnectivity.computeIfAbsent(groupOfElementsBreakingConnectivity, key -> new LinkedList<>()).add(contingency);
+            }
+        }
+    }
+
+    private static Set<ComputedContingencyElement> getGroupOfElementsBreakingConnectivity(LfNetwork lfNetwork, DenseMatrix contingenciesStates,
+                                                                                          Collection<ComputedContingencyElement> contingencyElements,
+                                                                                          EquationSystem<DcVariableType, DcEquationType> equationSystem) {
+        // use a sensitivity-criterion to detect the loss of connectivity after a contingency
+        // we consider a +1 -1 on a line, and we observe the sensitivity of these injections on the other contingency elements
+        // if the sum of the sensitivities (in absolute value) is 1, it means that all the flow is going through the lines with a non-zero sensitivity
+        // thus, losing these lines will lose the connectivity
+        Set<ComputedContingencyElement> groupOfElementsBreakingConnectivity = new LinkedHashSet<>();
+        for (ComputedContingencyElement element : contingencyElements) {
+            Set<ComputedContingencyElement> responsibleElements = new LinkedHashSet<>();
+            double sum = 0d;
+            for (ComputedContingencyElement element2 : contingencyElements) {
+                LfBranch branch = lfNetwork.getBranchById(element2.getElement().getId());
+                ClosedBranchSide1DcFlowEquationTerm p = equationSystem.getEquationTerm(ElementType.BRANCH, branch.getNum(), ClosedBranchSide1DcFlowEquationTerm.class);
+                double value = Math.abs(p.calculateSensi(contingenciesStates, element.getContingencyIndex()));
+                if (value > CONNECTIVITY_LOSS_THRESHOLD) {
+                    responsibleElements.add(element2);
+                }
+                sum += value;
+            }
+            if (sum > 1d - CONNECTIVITY_LOSS_THRESHOLD) {
+                // all lines that have a non-0 sensitivity associated to "element" breaks the connectivity
+                groupOfElementsBreakingConnectivity.addAll(responsibleElements);
+            }
+        }
+        return groupOfElementsBreakingConnectivity;
+    }
+
+    private static List<ConnectivityAnalysisResult> computeConnectivityData(LfNetwork lfNetwork, SensitivityFactorHolder<DcVariableType, DcEquationType> factorHolder,
+                                                                            Map<Set<ComputedContingencyElement>, List<PropagatedContingency>> contingenciesByGroupOfElementsBreakingConnectivity,
+                                                                            List<PropagatedContingency> nonLosingConnectivityContingencies, SensitivityResultWriter resultWriter) {
+        if (contingenciesByGroupOfElementsBreakingConnectivity.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        Map<Set<ComputedContingencyElement>, ConnectivityAnalysisResult> connectivityAnalysisResults = new LinkedHashMap<>();
+
+        GraphConnectivity<LfBus, LfBranch> connectivity = lfNetwork.getConnectivity();
+        for (Map.Entry<Set<ComputedContingencyElement>, List<PropagatedContingency>> e : contingenciesByGroupOfElementsBreakingConnectivity.entrySet()) {
+            Set<ComputedContingencyElement> breakingConnectivityCandidates = e.getKey();
+            List<PropagatedContingency> contingencyList = e.getValue();
+
+            Set<ComputedContingencyElement> breakingConnectivityElements;
+            connectivity.startTemporaryChanges();
+            try {
+                ComputedContingencyElement.applyToConnectivity(lfNetwork, connectivity, breakingConnectivityCandidates);
+                // filter the branches that really impacts connectivity
+                breakingConnectivityElements = breakingConnectivityCandidates.stream()
+                        .filter(element -> isBreakingConnectivity(connectivity, element))
+                        .collect(Collectors.toCollection(LinkedHashSet::new));
+            } finally {
+                connectivity.undoTemporaryChanges();
+            }
+
+            if (breakingConnectivityElements.isEmpty()) {
+                // we did not break any connectivity
+                nonLosingConnectivityContingencies.addAll(contingencyList);
+            } else {
+                // only compute for factors that have to be computed for this contingency lost
+                List<String> contingenciesIds = contingencyList.stream().map(contingency -> contingency.getContingency().getId()).collect(Collectors.toList());
+
+                List<LfSensitivityFactor<DcVariableType, DcEquationType>> lfFactors = factorHolder.getFactorsForContingencies(contingenciesIds);
+                if (!lfFactors.isEmpty()) {
+                    connectivity.startTemporaryChanges();
+                    try {
+                        ComputedContingencyElement.applyToConnectivity(lfNetwork, connectivity, breakingConnectivityElements);
+                        ConnectivityAnalysisResult connectivityAnalysisResult = connectivityAnalysisResults.computeIfAbsent(breakingConnectivityElements, k -> {
+                            Set<String> elementsToReconnect = computeElementsToReconnect(connectivity, breakingConnectivityElements);
+                            return new ConnectivityAnalysisResult(elementsToReconnect, connectivity, lfNetwork);
+                        });
+                        connectivityAnalysisResult.getContingencies().addAll(contingencyList);
+                    } finally {
+                        connectivity.undoTemporaryChanges();
+                    }
+                } else {
+                    // write contingency status
+                    for (PropagatedContingency propagatedContingency : contingencyList) {
+                        resultWriter.writeContingencyStatus(propagatedContingency.getIndex(), SensitivityAnalysisResult.Status.SUCCESS);
+                    }
+                }
+            }
+        }
+        return new ArrayList<>(connectivityAnalysisResults.values());
+    }
+
+    private static boolean isBreakingConnectivity(GraphConnectivity<LfBus, LfBranch> connectivity, ComputedContingencyElement element) {
+        LfBranch lfBranch = element.getLfBranch();
+        return connectivity.getComponentNumber(lfBranch.getBus1()) != connectivity.getComponentNumber(lfBranch.getBus2());
+    }
+
+    /**
+     * Given the elements breaking the connectivity, extract the minimum number of elements which reconnect all connected components together
+     */
+    private static Set<String> computeElementsToReconnect(GraphConnectivity<LfBus, LfBranch> connectivity, Set<DcSensitivityAnalysis.ComputedContingencyElement> breakingConnectivityElements) {
+        Set<String> elementsToReconnect = new LinkedHashSet<>();
+
+        // We suppose we're reconnecting one by one each element breaking connectivity.
+        // At each step we look if the reconnection was needed on the connectivity level by maintaining a list of grouped connected components.
+        List<Set<Integer>> reconnectedCc = new ArrayList<>();
+        for (DcSensitivityAnalysis.ComputedContingencyElement element : breakingConnectivityElements) {
+            int cc1 = connectivity.getComponentNumber(element.getLfBranch().getBus1());
+            int cc2 = connectivity.getComponentNumber(element.getLfBranch().getBus2());
+
+            Set<Integer> recCc1 = reconnectedCc.stream().filter(s -> s.contains(cc1)).findFirst().orElseGet(() -> new HashSet<>(List.of(cc1)));
+            Set<Integer> recCc2 = reconnectedCc.stream().filter(s -> s.contains(cc2)).findFirst().orElseGet(() -> Set.of(cc2));
+            if (recCc1 != recCc2) {
+                // cc1 and cc2 are still separated:
+                // - mark the element as needed to reconnect all connected components together
+                // - update the list of grouped connected components
+                elementsToReconnect.add(element.getElement().getId());
+                reconnectedCc.remove(recCc2);
+                if (recCc1.size() == 1) {
+                    // adding the new set (the list of grouped connected components is not initialized with the singleton sets)
+                    reconnectedCc.add(recCc1);
+                }
+                recCc1.addAll(recCc2);
+            }
+        }
+
+        if (reconnectedCc.size() != 1 || reconnectedCc.get(0).size() != connectivity.getNbConnectedComponents()) {
+            LOGGER.error("Elements to reconnect computed do not reconnect all connected components together");
+        }
+
+        return elementsToReconnect;
+    }
+
+    /**
+     * Calculate sensitivity values for a post-contingency state.
+     * When a contingency involves the loss of a load or a generator, the slack distribution could changed
+     * or the sensitivity factors in case of GLSK.
+     */
+    private void calculateContingencySensitivityValues(PropagatedContingency contingency, SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups, DenseMatrix factorStates, DenseMatrix contingenciesStates,
+                                                       DenseMatrix flowStates, Collection<ComputedContingencyElement> contingencyElements, SensitivityResultWriter resultWriter,
+                                                       DcLoadFlowContext loadFlowContext, OpenLoadFlowParameters lfParametersExt,
+                                                       SensitivityFactorHolder<DcVariableType, DcEquationType> factorHolder, List<ParticipatingElement> participatingElements,
+                                                       DisabledNetwork disabledNetwork, ReportNode reportNode) {
+        List<LfSensitivityFactor<DcVariableType, DcEquationType>> factors = factorHolder.getFactorsForContingency(contingency.getContingency().getId());
+        if (contingency.getGeneratorIdsToLose().isEmpty() && contingency.getLoadIdsToLoose().isEmpty()) {
+            calculateSensitivityValues(loadFlowContext, factors, factorStates, contingenciesStates, flowStates, contingencyElements,
+                    contingency, resultWriter, disabledNetwork);
+            // write contingency status
+            if (contingency.hasNoImpact()) {
+                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.NO_IMPACT);
+            } else {
+                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.SUCCESS);
+            }
+        } else {
+            // if we have a contingency including the loss of a DC line or a generator or a load
+            // save base state for later restoration after each contingency
+            LfNetwork lfNetwork = loadFlowContext.getNetwork();
+            DcLoadFlowParameters lfParameters = loadFlowContext.getParameters();
+            NetworkState networkState = NetworkState.save(lfNetwork);
+            LfContingency lfContingency = contingency.toLfContingency(lfNetwork).orElse(null);
+            DenseMatrix newFactorStates = factorStates;
+            List<ParticipatingElement> newParticipatingElements = participatingElements;
+            boolean participatingElementsChanged = false;
+            boolean rhsChanged = false;
+            if (lfContingency != null) {
+                lfContingency.apply(lfParameters.getBalanceType());
+                participatingElementsChanged = isDistributedSlackOnGenerators(lfParameters) && !contingency.getGeneratorIdsToLose().isEmpty()
+                        || isDistributedSlackOnLoads(lfParameters) && !contingency.getLoadIdsToLoose().isEmpty();
+                if (factorGroups.hasMultiVariables()) {
+                    Set<LfBus> impactedBuses = lfContingency.getLoadAndGeneratorBuses();
+                    rhsChanged = rescaleGlsk(factorGroups, impactedBuses);
+                }
+                if (participatingElementsChanged) {
+                    if (isDistributedSlackOnGenerators(lfParameters)) {
+                        // deep copy of participatingElements, removing the participating LfGeneratorImpl whose targetP has been set to 0
+                        Set<LfGenerator> participatingGeneratorsToRemove = lfContingency.getLostGenerators();
+                        newParticipatingElements = participatingElements.stream()
+                                .filter(participatingElement -> !participatingGeneratorsToRemove.contains(participatingElement.getElement()))
+                                .map(participatingElement -> new ParticipatingElement(participatingElement.getElement(), participatingElement.getFactor()))
+                                .collect(Collectors.toList());
+                        normalizeParticipationFactors(newParticipatingElements);
+                    } else { // slack distribution on loads
+                        newParticipatingElements = getParticipatingElements(lfNetwork.getBuses(), lfParameters.getBalanceType(), lfParametersExt);
+                    }
+                }
+                if (participatingElementsChanged || rhsChanged) {
+                    newFactorStates = calculateFactorStates(loadFlowContext, factorGroups, newParticipatingElements);
+                }
+                // write contingency status
+                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.SUCCESS);
+            } else {
+                // write contingency status
+                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.NO_IMPACT);
+            }
+
+            DenseMatrix newFlowStates = calculateActivePowerFlows(loadFlowContext, factors,
+                    newParticipatingElements, disabledNetwork, reportNode);
+
+            calculateSensitivityValues(loadFlowContext, factors, newFactorStates, contingenciesStates, newFlowStates, contingencyElements,
+                    contingency, resultWriter, disabledNetwork);
+
+            networkState.restore();
+            if (participatingElementsChanged || rhsChanged) {
+                setBaseCaseSensitivityValues(factorGroups, factorStates);
+            }
+        }
+    }
+
+    private void calculateSensitivityValuesForContingencyList(DcLoadFlowContext loadFlowContext, OpenLoadFlowParameters lfParametersExt, SensitivityFactorHolder<DcVariableType, DcEquationType> validFactorHolder,
+                                                              SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups, DenseMatrix factorState, DenseMatrix contingenciesStates, DenseMatrix flowStates,
+                                                              Collection<PropagatedContingency> contingencies, Map<String, ComputedContingencyElement> contingencyElementByBranch,
+                                                              Set<LfBus> disabledBuses, List<ParticipatingElement> participatingElements, Set<String> elementsToReconnect,
+                                                              SensitivityResultWriter resultWriter, ReportNode reportNode, Set<LfBranch> partialDisabledBranches) {
+        DenseMatrix modifiedFlowStates = flowStates;
+
+        PhaseTapChangerContingenciesIndexing phaseTapChangerContingenciesIndexing = new PhaseTapChangerContingenciesIndexing(contingencies, contingencyElementByBranch, elementsToReconnect);
+
+        var lfNetwork = loadFlowContext.getNetwork();
+
+        // compute contingencies without loss of phase tap changer
+        // first we compute the ones without loss of phase tap changers (because we reuse the load flows from the pre contingency network for all of them)
+        for (PropagatedContingency contingency : phaseTapChangerContingenciesIndexing.getContingenciesWithoutPhaseTapChangerLoss()) {
+            Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
+                    .filter(element -> !elementsToReconnect.contains(element))
+                    .map(contingencyElementByBranch::get)
+                    .collect(Collectors.toList());
+
+            Set<LfBranch> disabledBranches = contingency.getBranchIdsToOpen().keySet().stream().map(lfNetwork::getBranchById).collect(Collectors.toSet());
+            disabledBranches.addAll(partialDisabledBranches);
+
+            calculateContingencySensitivityValues(contingency, factorGroups, factorState, contingenciesStates, modifiedFlowStates, contingencyElements, resultWriter,
+                    loadFlowContext, lfParametersExt, validFactorHolder, participatingElements,
+                    new DisabledNetwork(disabledBuses, disabledBranches), reportNode);
+        }
+
+        // then we compute the ones involving the loss of a phase tap changer (because we need to recompute the load flows)
+        for (Map.Entry<Set<LfBranch>, Collection<PropagatedContingency>> e : phaseTapChangerContingenciesIndexing.getContingenciesIndexedByPhaseTapChangers().entrySet()) {
+            Set<LfBranch> disabledPhaseTapChangers = e.getKey();
+            Collection<PropagatedContingency> propagatedContingencies = e.getValue();
+            List<String> contingenciesIds = propagatedContingencies.stream()
+                    .map(c -> c.getContingency().getId())
+                    .collect(Collectors.toList());
+            List<LfSensitivityFactor<DcVariableType, DcEquationType>> lfFactors = validFactorHolder.getFactorsForContingencies(contingenciesIds);
+            if (!lfFactors.isEmpty()) {
+                modifiedFlowStates = calculateActivePowerFlows(loadFlowContext, lfFactors, participatingElements, new DisabledNetwork(disabledBuses, disabledPhaseTapChangers), reportNode);
+            }
+            for (PropagatedContingency contingency : propagatedContingencies) {
+                Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
+                        .filter(element -> !elementsToReconnect.contains(element))
+                        .map(contingencyElementByBranch::get)
+                        .collect(Collectors.toList());
+
+                Set<LfBranch> disabledBranches = contingency.getBranchIdsToOpen().keySet().stream().map(lfNetwork::getBranchById).collect(Collectors.toSet());
+                disabledBranches.addAll(partialDisabledBranches);
+
+                calculateContingencySensitivityValues(contingency, factorGroups, factorState, contingenciesStates, modifiedFlowStates, contingencyElements, resultWriter,
+                        loadFlowContext, lfParametersExt, validFactorHolder, participatingElements,
+                        new DisabledNetwork(disabledBuses, disabledBranches), reportNode);
+            }
+        }
+    }
+
+    private void processContingenciesBreakingConnectivity(ConnectivityAnalysisResult connectivityAnalysisResult, DcLoadFlowContext loadFlowContext,
+                                                          LoadFlowParameters lfParameters, OpenLoadFlowParameters lfParametersExt,
+                                                          SensitivityFactorHolder<DcVariableType, DcEquationType> validFactorHolder,
+                                                          SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups,
+                                                          List<ParticipatingElement> participatingElements,
+                                                          Map<String, ComputedContingencyElement> contingencyElementByBranch,
+                                                          DenseMatrix flowStates, DenseMatrix factorsStates, DenseMatrix contingenciesStates,
+                                                          SensitivityResultWriter resultWriter,
+                                                          ReportNode reportNode) {
+        DenseMatrix modifiedFlowStates = flowStates;
+
+        List<String> contingenciesIds = connectivityAnalysisResult.getContingencies().stream().map(c -> c.getContingency().getId()).collect(Collectors.toList());
+        List<LfSensitivityFactor<DcVariableType, DcEquationType>> lfFactorsForContingencies = validFactorHolder.getFactorsForContingencies(contingenciesIds);
+
+        Set<LfBus> disabledBuses = connectivityAnalysisResult.getDisabledBuses();
+        Set<LfBranch> partialDisabledBranches = connectivityAnalysisResult.getPartialDisabledBranches();
+
+        // as we are processing contingencies with connectivity break, we have to reset active power flow of a hvdc line
+        // if one bus of the line is lost.
+        for (LfHvdc hvdc : loadFlowContext.getNetwork().getHvdcs()) {
+            if (Networks.isIsolatedBusForHvdc(hvdc.getBus1(), disabledBuses) ^ Networks.isIsolatedBusForHvdc(hvdc.getBus2(), disabledBuses)) {
+                connectivityAnalysisResult.getContingencies().forEach(contingency -> {
+                    contingency.getGeneratorIdsToLose().add(hvdc.getConverterStation1().getId());
+                    contingency.getGeneratorIdsToLose().add(hvdc.getConverterStation2().getId());
+                });
+            }
+        }
+
+        // null and unused if slack bus is not distributed
+        List<ParticipatingElement> participatingElementsForThisConnectivity = participatingElements;
+        boolean rhsChanged = false; // true if the disabled buses change the slack distribution, or the GLSK
+        DenseMatrix factorStateForThisConnectivity = factorsStates;
+        if (lfParameters.isDistributedSlack()) {
+            rhsChanged = participatingElements.stream().anyMatch(element -> disabledBuses.contains(element.getLfBus()));
+        }
+        if (factorGroups.hasMultiVariables()) {
+            // some elements of the GLSK may not be in the connected component anymore, we recompute the injections
+            rhsChanged |= rescaleGlsk(factorGroups, disabledBuses);
+        }
+
+        // we need to recompute the factor states because the connectivity changed
+        if (rhsChanged) {
+            participatingElementsForThisConnectivity = lfParameters.isDistributedSlack()
+                    ? getParticipatingElements(connectivityAnalysisResult.getSlackConnectedComponent(), lfParameters.getBalanceType(), lfParametersExt) // will also be used to recompute the loadflow
+                    : Collections.emptyList();
+
+            factorStateForThisConnectivity = calculateFactorStates(loadFlowContext, factorGroups, participatingElementsForThisConnectivity);
+        }
+
+        if (!lfFactorsForContingencies.isEmpty()) {
+            modifiedFlowStates = calculateActivePowerFlows(loadFlowContext, lfFactorsForContingencies,
+                    participatingElementsForThisConnectivity, new DisabledNetwork(disabledBuses, Collections.emptySet()),
+                    reportNode);
+        }
+
+        calculateSensitivityValuesForContingencyList(loadFlowContext, lfParametersExt,
+                validFactorHolder, factorGroups, factorStateForThisConnectivity, contingenciesStates, modifiedFlowStates,
+                connectivityAnalysisResult.getContingencies(), contingencyElementByBranch, disabledBuses, participatingElementsForThisConnectivity,
+                connectivityAnalysisResult.getElementsToReconnect(), resultWriter, reportNode, partialDisabledBranches);
+
+        if (rhsChanged) {
+            setBaseCaseSensitivityValues(factorGroups, factorsStates); // we modified the rhs, we need to restore previous state
+        }
+    }
+
+    private static Map<String, ComputedContingencyElement> createContingencyElementsIndexByBranchId(List<PropagatedContingency> contingencies,
+                                                                                                    LfNetwork lfNetwork, EquationSystem<DcVariableType, DcEquationType> equationSystem) {
+        Map<String, ComputedContingencyElement> contingencyElementByBranch =
+                contingencies.stream()
+                        .flatMap(contingency -> contingency.getBranchIdsToOpen().keySet().stream())
+                        .map(branch -> new ComputedContingencyElement(new BranchContingency(branch), lfNetwork, equationSystem))
+                        .filter(element -> element.getLfBranchEquation() != null)
+                        .collect(Collectors.toMap(
+                            computedContingencyElement -> computedContingencyElement.getElement().getId(),
+                            computedContingencyElement -> computedContingencyElement,
+                            (existing, replacement) -> existing,
+                            LinkedHashMap::new
+                        ));
+        ComputedContingencyElement.setContingencyIndexes(contingencyElementByBranch.values());
+        return contingencyElementByBranch;
     }
 
     protected void cleanContingencies(LfNetwork lfNetwork, List<PropagatedContingency> contingencies) {
@@ -352,10 +937,10 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
 
         // create the network (we only manage main connected component)
         SlackBusSelector slackBusSelector = SlackBusSelector.fromMode(lfParametersExt.getSlackBusSelectionMode(),
-                lfParametersExt.getSlackBusesIds(),
-                lfParametersExt.getPlausibleActivePowerLimit(),
-                lfParametersExt.getMostMeshedSlackBusSelectorMaxNominalVoltagePercentile(),
-                lfParametersExt.getSlackBusCountryFilter());
+                                                                      lfParametersExt.getSlackBusesIds(),
+                                                                      lfParametersExt.getPlausibleActivePowerLimit(),
+                                                                      lfParametersExt.getMostMeshedSlackBusSelectorMaxNominalVoltagePercentile(),
+                                                                      lfParametersExt.getSlackBusCountryFilter());
         if (lfParameters.isReadSlackBus()) {
             slackBusSelector = new NetworkSlackBusSelector(network, lfParametersExt.getSlackBusCountryFilter(), slackBusSelector);
         }
@@ -395,10 +980,10 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
 
             allLfFactors.stream()
                     .filter(lfFactor -> lfFactor.getFunctionType() != SensitivityFunctionType.BRANCH_ACTIVE_POWER_1
-                            && lfFactor.getFunctionType() != SensitivityFunctionType.BRANCH_ACTIVE_POWER_2
+                                && lfFactor.getFunctionType() != SensitivityFunctionType.BRANCH_ACTIVE_POWER_2
                             || lfFactor.getVariableType() != SensitivityVariableType.INJECTION_ACTIVE_POWER
-                            && lfFactor.getVariableType() != SensitivityVariableType.TRANSFORMER_PHASE
-                            && lfFactor.getVariableType() != SensitivityVariableType.HVDC_LINE_ACTIVE_POWER)
+                                && lfFactor.getVariableType() != SensitivityVariableType.TRANSFORMER_PHASE
+                                && lfFactor.getVariableType() != SensitivityVariableType.HVDC_LINE_ACTIVE_POWER)
                     .findFirst()
                     .ifPresent(ignored -> {
                         throw new PowsyblException("Only variables of type TRANSFORMER_PHASE, INJECTION_ACTIVE_POWER and HVDC_LINE_ACTIVE_POWER, and functions of type BRANCH_ACTIVE_POWER_1 and BRANCH_ACTIVE_POWER_2 are yet supported in DC");
@@ -423,7 +1008,7 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
                 DcLoadFlowEngine.initStateVector(lfNetwork, loadFlowContext.getEquationSystem(), voltageInitializer);
 
                 // index factors by variable group to compute the minimal number of states
-                SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups = createFactorGroups(validLfFactors.stream().filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID).toList());
+                SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups = createFactorGroups(validLfFactors.stream().filter(factor -> factor.getStatus() == LfSensitivityFactor.Status.VALID).collect(Collectors.toList()));
 
                 // compute the participation for each injection factor (+1 on the injection and then -participation factor on all
                 // buses that contain elements participating to slack distribution)
@@ -431,254 +1016,60 @@ public class DcSensitivityAnalysis extends AbstractSensitivityAnalysis<DcVariabl
                         ? getParticipatingElements(lfNetwork.getBuses(), lfParameters.getBalanceType(), lfParametersExt)
                         : Collections.emptyList();
 
-                // compute states with +1 -1 to model the contingencies and run connectivity analysis
-                ConnectivityBreakAnalysis.ConnectivityBreakAnalysisResults connectivityData = ConnectivityBreakAnalysis.run(loadFlowContext, contingencies);
+                // index contingency elements by branch id
+                Map<String, ComputedContingencyElement> contingencyElementByBranch = createContingencyElementsIndexByBranchId(contingencies, lfNetwork, loadFlowContext.getEquationSystem());
 
-                // engine to compute post-contingency states values for sensitivities calculations
-                WoodburyEngine engine = new WoodburyEngine();
+                // run DC load on pre-contingency network
+                DenseMatrix flowStates = calculateActivePowerFlows(loadFlowContext, validLfFactors, participatingElements, new DisabledNetwork(), reportNode);
 
-                // pre-contingency flow states
-                DenseMatrix preContingencyFlowStates = getPreContingencyFlowRhs(loadFlowContext, participatingElements, new DisabledNetwork());
-                solve(preContingencyFlowStates, loadFlowContext.getJacobianMatrix(), reportNode);
+                // compute the pre-contingency sensitivity values
+                DenseMatrix factorsStates = calculateFactorStates(loadFlowContext, factorGroups, participatingElements);
 
-                // set function reference values of the factors
-                setFunctionReference(validLfFactors, preContingencyFlowStates);
+                // calculate sensitivity values for pre-contingency network
+                calculateSensitivityValues(loadFlowContext, validFactorHolder.getFactorsForBaseNetwork(), factorsStates, null, flowStates,
+                        Collections.emptySet(), null, resultWriter, new DisabledNetwork());
 
-                WoodburyEngineInputReader flowInputReader = new AbstractWoodburyEngineInputReader() {
-                    Set<LfBranch> getDisabledPhaseTapChangers(PropagatedContingency contingency, Set<String> elementsToReconnect) {
-                        return contingency.getBranchIdsToOpen().keySet().stream()
-                                .filter(element -> !elementsToReconnect.contains(element))
-                                .map(connectivityData.contingencyElementByBranch()::get)
-                                .map(ComputedContingencyElement::getLfBranch)
-                                .filter(LfBranch::hasPhaseControllerCapability)
-                                .collect(Collectors.toSet());
-                    }
+                // compute states with +1 -1 to model the contingencies
+                DenseMatrix contingenciesStates = calculateContingenciesStates(loadFlowContext, contingencyElementByBranch);
 
-                    Optional<DenseMatrix> getPreContingencyFlowRhsOverride(PropagatedContingency contingency, Set<LfBus> disabledBuses, Set<LfBranch> partialDisabledBranches,
-                                                                           Set<String> elementsToReconnect, List<ParticipatingElement> participatingElements) {
-                        if (!contingency.getGeneratorIdsToLose().isEmpty() || !contingency.getLoadIdsToLoose().isEmpty()) {
-                            Set<LfBranch> disabledBranches = contingency.getBranchIdsToOpen().keySet().stream().map(lfNetwork::getBranchById).collect(Collectors.toSet());
-                            disabledBranches.addAll(partialDisabledBranches);
-                            DisabledNetwork disabledNetwork = new DisabledNetwork(disabledBuses, disabledBranches);
+                // connectivity analysis by contingency
+                // we have to compute sensitivities and reference functions in a different way depending on either or not the contingency breaks connectivity
+                // so, we will index contingencies by a list of branch that may break connectivity
+                // for example, if in the network, loosing line L1 breaks connectivity, and loosing L2 and L3 together breaks connectivity,
+                // the index would be: {L1, L2, L3}
+                // a contingency involving a phase tap changer loss has to be processed separately
+                List<PropagatedContingency> nonBreakingConnectivityContingencies = new ArrayList<>();
+                Map<Set<ComputedContingencyElement>, List<PropagatedContingency>> contingenciesByGroupOfElementsPotentiallyBreakingConnectivity = new LinkedHashMap<>();
 
-                            // if we have a contingency including the loss of a DC line or a generator or a load
-                            // save base state for later restoration
-                            LfNetwork lfNetwork = loadFlowContext.getNetwork();
-                            NetworkState networkState = NetworkState.save(lfNetwork);
-                            LfContingency lfContingency = contingency.toLfContingency(lfNetwork).orElse(null);
-                            List<ParticipatingElement> newParticipatingElements = participatingElements;
-                            if (lfContingency != null) {
-                                lfContingency.apply(lfParameters.getBalanceType());
-                                // if participating elements are changed
-                                if (isDistributedSlackOnGenerators(loadFlowContext.getParameters()) && !contingency.getGeneratorIdsToLose().isEmpty()
-                                        || isDistributedSlackOnLoads(loadFlowContext.getParameters()) && !contingency.getLoadIdsToLoose().isEmpty()) {
-                                    newParticipatingElements = getNewNormalizedParticipationFactors(loadFlowContext, lfParametersExt, lfContingency, newParticipatingElements);
-                                }
-                            }
-                            DenseMatrix preContingencyFlowRhsOverride = getPreContingencyFlowRhs(loadFlowContext, newParticipatingElements, disabledNetwork);
-                            networkState.restore();
-                            return Optional.of(preContingencyFlowRhsOverride);
-                        } else {
-                            // in case a phase tap changer is lost, flow rhs must be updated
-                            Set<LfBranch> disabledPhaseTapChangers = getDisabledPhaseTapChangers(contingency, elementsToReconnect);
-                            if (!disabledPhaseTapChangers.isEmpty()) {
-                                return Optional.of(getPreContingencyFlowRhs(loadFlowContext, participatingElements,
-                                        new DisabledNetwork(disabledBuses, disabledPhaseTapChangers)));
-                            } else {
-                                // no override is needed
-                                return Optional.empty();
-                            }
-                        }
-                    }
+                // this first method based on sensitivity criteria is able to detect some contingencies that do not break
+                // connectivity and other contingencies that potentially break connectivity
+                detectPotentialConnectivityBreak(lfNetwork, contingenciesStates, contingencies, contingencyElementByBranch, loadFlowContext.getEquationSystem(),
+                        nonBreakingConnectivityContingencies, contingenciesByGroupOfElementsPotentiallyBreakingConnectivity);
+                LOGGER.info("After sensitivity based connectivity analysis, {} contingencies do not break connectivity, {} contingencies potentially break connectivity",
+                        nonBreakingConnectivityContingencies.size(), contingenciesByGroupOfElementsPotentiallyBreakingConnectivity.values().stream().mapToInt(List::size).count());
 
-                    @Override
-                    public void process(Handler handler) {
+                // this second method process all contingencies that potentially break connectivity and using graph algorithms
+                // find remaining contingencies that do not break connectivity
+                List<ConnectivityAnalysisResult> connectivityAnalysisResults
+                        = computeConnectivityData(lfNetwork, validFactorHolder, contingenciesByGroupOfElementsPotentiallyBreakingConnectivity, nonBreakingConnectivityContingencies, resultWriter);
+                LOGGER.info("After graph based connectivity analysis, {} contingencies do not break connectivity, {} contingencies break connectivity",
+                        nonBreakingConnectivityContingencies.size(), connectivityAnalysisResults.stream().mapToInt(results -> results.getContingencies().size()).count());
 
-                        Map<String, ComputedContingencyElement> contingencyElementByBranch = connectivityData.contingencyElementByBranch();
-                        for (ConnectivityBreakAnalysis.ConnectivityAnalysisResult connectivityAnalysisResult : connectivityData.connectivityAnalysisResults()) {
-                            List<ParticipatingElement> participatingElementsForThisConnectivity = participatingElements;
-                            Set<LfBus> disabledBuses = connectivityAnalysisResult.getDisabledBuses();
+                LOGGER.info("Processing contingencies with no connectivity break");
 
-                            // as we are processing contingencies with connectivity break, we have to reset active power flow of a hvdc line
-                            // if one bus of the line is lost.
-                            processHvdcLinesWithDisconnection(loadFlowContext, disabledBuses, connectivityAnalysisResult);
+                // process contingencies with no connectivity break
+                calculateSensitivityValuesForContingencyList(loadFlowContext, lfParametersExt, validFactorHolder, factorGroups,
+                        factorsStates, contingenciesStates, flowStates, nonBreakingConnectivityContingencies, contingencyElementByBranch,
+                        Collections.emptySet(), participatingElements, Collections.emptySet(), resultWriter, reportNode, Collections.emptySet());
 
-                            // we need to recompute the rhs because the connectivity changed
-                            boolean rhsChanged = hasRhsChangedDueToDisabledSlackBus(lfParameters, disabledBuses, participatingElementsForThisConnectivity);
-                            if (rhsChanged) {
-                                participatingElementsForThisConnectivity = new ArrayList<>(lfParameters.isDistributedSlack()
-                                        ? getParticipatingElements(connectivityAnalysisResult.getSlackConnectedComponent(), lfParameters.getBalanceType(), lfParametersExt) // will also be used to recompute the loadflow
-                                        : Collections.emptyList());
-                            }
+                LOGGER.info("Processing contingencies with connectivity break");
 
-                            DisabledNetwork disabledNetwork = new DisabledNetwork(disabledBuses, Collections.emptySet());
-                            DenseMatrix preContingencyStatesOverrideConnectivityBreak = getPreContingencyFlowRhs(loadFlowContext, participatingElementsForThisConnectivity, disabledNetwork);
-                            // compute pre-contingency states values override
-                            solve(preContingencyStatesOverrideConnectivityBreak, loadFlowContext.getJacobianMatrix(), reportNode);
-
-                            Set<String> elementsToReconnect = connectivityAnalysisResult.getElementsToReconnect();
-                            Set<LfBranch> partialDisabledBranches = connectivityAnalysisResult.getPartialDisabledBranches();
-                            for (PropagatedContingency contingency : connectivityAnalysisResult.getContingencies()) {
-                                Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
-                                        .filter(element -> !elementsToReconnect.contains(element))
-                                        .map(contingencyElementByBranch::get)
-                                        .toList();
-                                Optional<DenseMatrix> preContingencyStatesOverride = getPreContingencyFlowRhsOverride(contingency, disabledBuses, partialDisabledBranches,
-                                        elementsToReconnect, participatingElementsForThisConnectivity);
-                                preContingencyStatesOverride.ifPresentOrElse(override -> {
-                                    // compute pre-contingency states values override
-                                    solve(override, loadFlowContext.getJacobianMatrix(), reportNode);
-                                    handler.onContingency(contingency, contingencyElements, override);
-                                }, () -> handler.onContingency(contingency, contingencyElements, preContingencyStatesOverrideConnectivityBreak));
-                            }
-                        }
-
-                        for (PropagatedContingency contingency : connectivityData.nonBreakingConnectivityContingencies()) {
-                            Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
-                                    .map(contingencyElementByBranch::get)
-                                    .toList();
-                            Optional<DenseMatrix> preContingencyStatesOverride = getPreContingencyFlowRhsOverride(contingency, Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), participatingElements);
-                            preContingencyStatesOverride.ifPresentOrElse(override -> {
-                                // compute pre-contingency states values override
-                                solve(override, loadFlowContext.getJacobianMatrix(), reportNode);
-                                handler.onContingency(contingency, contingencyElements, override);
-                            }, () -> handler.onContingency(contingency, contingencyElements, preContingencyFlowStates));
-                        }
-                    }
-                };
-
-                // compute post-contingency flow states
-                Map<PropagatedContingency, DenseMatrix> flowResult = engine.run(loadFlowContext, flowInputReader, connectivityData.contingenciesStates());
-
-                // storage of disabled network by propagated contingencies, for sensitivity calculation
-                HashMap<PropagatedContingency, DisabledNetwork> disabledNetworkByPropagatedContingency = new HashMap<>();
-
-                // pre-contingency injection states
-                DenseMatrix preContingencyInjectionStates = getPreContingencyInjectionRhs(loadFlowContext, factorGroups, participatingElements);
-                solve(preContingencyInjectionStates, loadFlowContext.getJacobianMatrix(), reportNode);
-
-                // set base case values of the factors
-                setBaseCaseSensitivityValues(factorGroups, preContingencyInjectionStates);
-
-                WoodburyEngineInputReader injectionInputReader = new AbstractWoodburyEngineInputReader() {
-
-                    /**
-                     * @return True if the disabled buses change the slack distribution or the GLSK. False otherwise.
-                     */
-                    boolean hasRhsChangedDueToConnectivityBreak(LoadFlowParameters lfParameters, SensitivityFactorGroupList<DcVariableType, DcEquationType> factorGroups,
-                                                                Set<LfBus> disabledBuses, List<ParticipatingElement> participatingElements) {
-                        boolean rhsChanged = hasRhsChangedDueToDisabledSlackBus(lfParameters, disabledBuses, participatingElements);
-                        if (factorGroups.hasMultiVariables()) {
-                            // some elements of the GLSK may not be in the connected component anymore, we recompute the injections
-                            rhsChanged |= rescaleGlsk(factorGroups, disabledBuses);
-                        }
-                        return rhsChanged;
-                    }
-
-                    Optional<DenseMatrix> getPreContingencyInjectionRhsOverride(PropagatedContingency contingency, Set<LfBus> disabledBuses, Set<LfBranch> partialDisabledBranches) {
-                        DenseMatrix preContingencyStatesOverride = null;
-
-                        Set<LfBranch> disabledBranches = contingency.getBranchIdsToOpen().keySet().stream().map(lfNetwork::getBranchById).collect(Collectors.toSet());
-                        disabledBranches.addAll(partialDisabledBranches);
-                        DisabledNetwork disabledNetwork = new DisabledNetwork(disabledBuses, disabledBranches);
-                        disabledNetworkByPropagatedContingency.put(contingency, disabledNetwork);
-
-                        if (contingency.getGeneratorIdsToLose().isEmpty() && contingency.getLoadIdsToLoose().isEmpty()) {
-                            resultWriter.writeContingencyStatus(contingency.getIndex(), contingency.hasNoImpact() ? SensitivityAnalysisResult.Status.NO_IMPACT : SensitivityAnalysisResult.Status.SUCCESS);
-                        } else {
-                            // if we have a contingency including the loss of a DC line or a generator or a load
-                            // save base state for later restoration after each contingency
-                            LfNetwork lfNetwork = loadFlowContext.getNetwork();
-                            NetworkState networkState = NetworkState.save(lfNetwork);
-                            LfContingency lfContingency = contingency.toLfContingency(lfNetwork).orElse(null);
-                            if (lfContingency != null) {
-                                DcLoadFlowParameters lfParameters = loadFlowContext.getParameters();
-                                lfContingency.apply(lfParameters.getBalanceType());
-                                List<ParticipatingElement> modifiedParticipatingElements = participatingElements;
-                                boolean rhsChanged = isDistributedSlackOnGenerators(loadFlowContext.getParameters()) && !contingency.getGeneratorIdsToLose().isEmpty()
-                                        || isDistributedSlackOnLoads(loadFlowContext.getParameters()) && !contingency.getLoadIdsToLoose().isEmpty();
-                                if (rhsChanged) {
-                                    modifiedParticipatingElements = getNewNormalizedParticipationFactors(loadFlowContext, lfParametersExt, lfContingency, modifiedParticipatingElements);
-                                }
-                                if (factorGroups.hasMultiVariables()) {
-                                    Set<LfBus> impactedBuses = lfContingency.getLoadAndGeneratorBuses();
-                                    rhsChanged |= DcSensitivityAnalysis.this.rescaleGlsk(factorGroups, impactedBuses);
-                                }
-                                if (rhsChanged) {
-                                    preContingencyStatesOverride = getPreContingencyInjectionRhs(loadFlowContext, factorGroups, modifiedParticipatingElements);
-                                }
-                                // write contingency status
-                                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.SUCCESS);
-                            } else {
-                                // write contingency status
-                                resultWriter.writeContingencyStatus(contingency.getIndex(), SensitivityAnalysisResult.Status.NO_IMPACT);
-                            }
-                            networkState.restore();
-                        }
-
-                        return Optional.ofNullable(preContingencyStatesOverride);
-                    }
-
-                    @Override
-                    public void process(Handler handler) {
-
-                        Map<String, ComputedContingencyElement> contingencyElementByBranch = connectivityData.contingencyElementByBranch();
-                        for (ConnectivityBreakAnalysis.ConnectivityAnalysisResult connectivityAnalysisResult : connectivityData.connectivityAnalysisResults()) {
-                            List<ParticipatingElement> participatingElementsForThisConnectivity = participatingElements;
-                            Set<LfBus> disabledBuses = connectivityAnalysisResult.getDisabledBuses();
-
-                            // as we are processing contingencies with connectivity break, we have to reset active power flow of a hvdc line
-                            // if one bus of the line is lost.
-                            processHvdcLinesWithDisconnection(loadFlowContext, disabledBuses, connectivityAnalysisResult);
-
-                            // we need to recompute the injection rhs because the connectivity changed
-                            DenseMatrix preContingencyStatesOverrideConnectivityBreak;
-                            boolean rhsChanged = hasRhsChangedDueToConnectivityBreak(lfParameters, factorGroups, disabledBuses, participatingElementsForThisConnectivity);
-                            if (rhsChanged) {
-                                participatingElementsForThisConnectivity = new ArrayList<>(lfParameters.isDistributedSlack()
-                                        ? DcSensitivityAnalysis.this.getParticipatingElements(connectivityAnalysisResult.getSlackConnectedComponent(), lfParameters.getBalanceType(), lfParametersExt) // will also be used to recompute the loadflow
-                                        : Collections.emptyList());
-                                preContingencyStatesOverrideConnectivityBreak = getPreContingencyInjectionRhs(loadFlowContext, factorGroups, participatingElementsForThisConnectivity);
-                                // compute pre-contingency states values override
-                                solve(preContingencyStatesOverrideConnectivityBreak, loadFlowContext.getJacobianMatrix(), reportNode);
-                            } else {
-                                preContingencyStatesOverrideConnectivityBreak = preContingencyInjectionStates;
-                            }
-
-                            Set<String> elementsToReconnect = connectivityAnalysisResult.getElementsToReconnect();
-                            for (PropagatedContingency contingency : connectivityAnalysisResult.getContingencies()) {
-                                Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
-                                        .filter(element -> !elementsToReconnect.contains(element))
-                                        .map(contingencyElementByBranch::get)
-                                        .toList();
-                                Optional<DenseMatrix> preContingencyInjectionStatesOverride = getPreContingencyInjectionRhsOverride(contingency, disabledBuses, connectivityAnalysisResult.getPartialDisabledBranches());
-                                preContingencyInjectionStatesOverride.ifPresentOrElse(override -> {
-                                    // compute pre-contingency states values override
-                                    solve(override, loadFlowContext.getJacobianMatrix(), reportNode);
-                                    handler.onContingency(contingency, contingencyElements, override);
-                                }, () -> handler.onContingency(contingency, contingencyElements, preContingencyStatesOverrideConnectivityBreak));
-                            }
-                        }
-
-                        for (PropagatedContingency contingency : connectivityData.nonBreakingConnectivityContingencies()) {
-                            Collection<ComputedContingencyElement> contingencyElements = contingency.getBranchIdsToOpen().keySet().stream()
-                                    .map(contingencyElementByBranch::get)
-                                    .toList();
-                            Optional<DenseMatrix> preContingencyStatesOverride = getPreContingencyInjectionRhsOverride(contingency, Collections.emptySet(), Collections.emptySet());
-                            preContingencyStatesOverride.ifPresentOrElse(override -> {
-                                // compute pre-contingency states values override
-                                solve(override, loadFlowContext.getJacobianMatrix(), reportNode);
-                                handler.onContingency(contingency, contingencyElements, override);
-                            }, () -> handler.onContingency(contingency, contingencyElements, preContingencyInjectionStates));
-                        }
-                    }
-                };
-
-                // compute pre- and post-contingency injection states
-                Map<PropagatedContingency, DenseMatrix> injectionResult = engine.run(loadFlowContext, injectionInputReader, connectivityData.contingenciesStates());
-
-                // compute the sensibilities with Woodbury post-contingency states, and computed disabledNetworks
-                calculateSensitivityValues(injectionResult, flowResult, disabledNetworkByPropagatedContingency, validFactorHolder.getAllFactors(), contingencies, resultWriter);
+                // process contingencies with connectivity break
+                for (ConnectivityAnalysisResult connectivityAnalysisResult : connectivityAnalysisResults) {
+                    processContingenciesBreakingConnectivity(connectivityAnalysisResult, loadFlowContext, lfParameters, lfParametersExt,
+                            validFactorHolder, factorGroups, participatingElements,
+                            contingencyElementByBranch, flowStates, factorsStates, contingenciesStates, resultWriter, reportNode);
+                }
             }
 
             stopwatch.stop();
